+++ demo/src/main/java/com/example/demo/CursoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface CursoRepository extends JpaRepository<Curso, BigDecimal> {
+public interface CursoRepository extends JpaRepository<Curso, Long> {
 }
