+++ demo/src/main/java/com/example/demo/PrestamosLibroRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface PrestamosLibroRepository extends JpaRepository<PrestamosLibro, BigDecimal> {
+public interface PrestamosLibroRepository extends JpaRepository<PrestamosLibro, Long> {
 }
